progress loader + rename
