+++ src/components/app/ActionCard/index.js
@@ -1,1 +1,1 @@
-export { default } from './DetailCard';
+export { default } from './ActionCard';
